feat: add conformance to common protocols
This adds conformance to the basic collection protocols, and updates the
initialisers to also reflect this (namely, dropping the initialiser
which just returns unallocated memory to be filled in later). We can
also drop the .toArray() and .toMultiArray() functions as they are no
longer needed.
 - Equatable
 - Hashable
 - CustomStringConvertible
 - CustomDebugStringConvertible
 - ExpressibleByArrayLiteral
 - MutableCollection
 - RandomAccessCollection
 - Encodable
 - Decodable

These do the expected thing and work over the Element type, although for
representations that do not include Box it would be more efficient to
run over the raw pointer. If we eventually provide this functionality it
should probably be via separate functions.

We still provide our own .map() function, shadowing the one provided via
Sequence, as that one does not enable vectorisation. More investigation
required to see if we could enable this via the standard interface.