fix: initialise/deinitialise of (boxed) array data
This is required for any class-based types (e.g. String) that we are
storing via Box. When first writing to the raw buffer we must initialise
the memory (retaining as strong reference to the value that is not owned
by us) and ensure that memory is deinitialised when the containing
MultiArray is deallocated (releasing said strong reference). Previously
we were not doing the latter, and when modifying elements (subscript
write) re-initialising that slot. Both of these would cause memory
leaks.